third_party/blink/web_tests/external/wpt/dom/nodes/moveBefore/tentative/selection-preserve-expected.txt

This is a testharness.js-based test.
[FAIL] moveBefore should not reset selection with preceding text
  assert_equals: expected Text node "Text" but got Text node "This text does not move"
[FAIL] moveBefore should not reset selection with preceding and following text
  assert_equals: expected Text node "Text" but got Text node "\n      "
Harness: the test ran to completion.
